factor CreatePlayer into own control, general refactor/code cleanup.
