+++ ASP.NET/HeroCommandReact/ClientApp/src/App.js
@@ -3,7 +3,6 @@ import { Route } from 'react-router';
 import { Layout } from './components/Layout';
 import { Home } from './components/Home';
 import { FetchData } from './components/FetchData';
-import { Counter } from './components/Counter';
 import { PlayerLogin } from './components/PlayerLogin';
 import { AdminControl } from './components/Admin';
 import { Gameplay } from './components/Gameplay';
@@ -17,11 +16,10 @@ export default class App extends Component {
     return (
         <Layout>
         <Route exact path='/' component={Home} />
-        <Route path='/counter' component={Counter} />
-            <Route path='/fetch-data' component={FetchData} />
-            <Route path='/player-login' component={PlayerLogin} />
-            <Route path='/admin' component={AdminControl} />
-            <Route path='/gameplay' component={Gameplay} />
+        <Route path='/fetch-data' component={FetchData} />
+        <Route path='/player-login' component={PlayerLogin} />
+        <Route path='/admin' component={AdminControl} />
+        <Route path='/gameplay' component={Gameplay} />
       </Layout>
     );
   }
