+++ ASP.NET/HeroCommandReact/ClientApp/src/components/Home.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
 
 export class Home extends Component {
   static displayName = Home.name;
@@ -12,6 +13,9 @@ export class Home extends Component {
             As a reward for completing missions, heroes gain skill while you gain coin and experience.
             As you gain experience, more heroes become available for hire and more missions become available to undertake.
         </p>
+        <p>
+            To get started, head on over to the <Link to='/player-login'>Player Login</Link> page.
+        </p>
         <p>Hero Command is single-page application serving data through a REST API, built with:</p>
         <ul>
           <li><a href='https://get.asp.net/'>ASP.NET Core</a> and <a href='https://msdn.microsoft.com/en-us/library/67ef8sbd.aspx'>C#</a> for cross-platform server-side code</li>
@@ -20,7 +24,7 @@ export class Home extends Component {
         </ul>
         <p>
             Hero Command <strong>is not</strong> meant to be a fully-realized, playable game. 
-            Rather it has been designed as a technical demo to show off its creator's technical capabilities.
+            Rather it has been designed as a technical demo to show off its creator's skills.
             If you would like to take a look at the source code for Hero Command, please feel free to visit over on <a href='https://github.com/cstilstra/HeroCommand'>Github</a>.
         </p>
       </div>
